Go off canvas (#22)
* Creates a system for opening up components in an off canvas

The off canvas is a UI interaction, similar to a modal, but allows
us to better maintain the context of what is going on by opening up
to the side of the content that is being interacted with on the page.
This adds a component, directive, and a service to manage the off
canvas UI.

* Abstracts away the off canvas animations

This pulls all the animations out of the component and into a shared
/animations directory. This give us a shared location to put animations
and cleans up their implementation.

* Add a close button and update layout to make content scrollable

* Adds code to the go-tester app.component to test off canvas

This implements a couple of examples in the app.component inside
the go-tester project. These changes will show how to implement
the service and give us a visual way to test the components.

* Update code based on feedback and best practices

This includes some sizeable code changes to further adhere to
TypeScript and Angular best practices. It also cleans up several
issues that were brought up in code review.