Fixes the way data is exposed to the go-toast component (#33)
There was an issue we were running into when attempting to compile
goponents due to the toast component template not having access
to the functions that are private inside the component.ts file.

This keeps those private, but changes the way we expose those to
the template using static variables. This will fix this issue and
is probably a little bit more performant.