Fixes issue with loader svg in Safari. Closes #82
There was an issue in the Safari browser where the loading svg
would look like three black circles. This was a result of there
being duplicate IDs within the svgs markup.

Generating unique IDs and applying the fill to the paths inline
based on those unique IDs fixes this problem.