Fixes issue with loader svg in Safari. Closes #82 (#93)
* Fixes issue with loader svg in Safari. Closes #82

There was an issue in the Safari browser where the loading svg
would look like three black circles. This was a result of there
being duplicate IDs within the svgs markup.

Generating unique IDs and applying the fill to the paths inline
based on those unique IDs fixes this problem.

* Implements new version of the loading svg

The previous version of the loading svg had some issues in specific
browsers becuase it was reliant on gradient filters. This redesigns
the svg to no longer use gradients and uses angular to do some of
the animations on the host element.

* Adds tests for go-loader component

* Updates go-tester to test loader

* Add boolean to @HostBinding animation