V1.1.0 (#102)
* Adds Side Navigation Module

* Bugfix for empty table column Fixes #53

* Feature Go Search

* Code review revisions

* Adds all files necessary for GoIconButtonComponent (#57)

* Adds all files necessary for GoIconButtonComponent

Creates a component for button with icons with no text.

* Add title attribute to go-icon-button

* Init commit of go-layout

* Added header

* Added go-side-nav to go-layout

* Clean up disjointed rebase

* Rebase style issues with go-icon-button

* Formatting fixes

* Fix failing component creation tests

* Fix a rebase

* Updated menu icon to go-icon-button

* Updates to search animations

Fixes animation errors in firefox

* Add map for z-indexes

* Fix where z-index is applied for search and off-canvas

* Fixes issue with loader svg in Safari. Closes #82 (#93)

* Fixes issue with loader svg in Safari. Closes #82

There was an issue in the Safari browser where the loading svg
would look like three black circles. This was a result of there
being duplicate IDs within the svgs markup.

Generating unique IDs and applying the fill to the paths inline
based on those unique IDs fixes this problem.

* Implements new version of the loading svg

The previous version of the loading svg had some issues in specific
browsers becuase it was reliant on gradient filters. This redesigns
the svg to no longer use gradients and uses angular to do some of
the animations on the host element.

* Adds tests for go-loader component

* Updates go-tester to test loader

* Add boolean to @HostBinding animation

* Bug ie edge cases (#91)

* Fixes IE and Edge bugs with components for v1.1

* Removes 'x' on search input

* Fix title attr and center alignment

* Implemented styles and functionality for a mobile layout (#94)

* [Feature] Go-Action-Sheet (#90)

* Added the ability to provide actions on a table in a toolbar at the top (#97)

* Renames title to panelContent to avoid conflicts with native functionality (#98)

We had an issue with the <go-panel> component where the title @input
was exposing things we didn't want it to because of its naming
collision with the native title attribute. Updating the name of said
attribute to panelContent fixes this issue.

* Fixes compilation errors in the navigation components due to type checking