Refactor accordion component (#129)
* Refactors the accordion component

This commit turns back on view encapsulation and refactors the
accordion component. The most notable changes are the removal of the
accordion template and reworking of all of the styles to correspond
with the components they apply to.

Closes #122, Closes #87

* Upgrade the jasmine module

* Refactors and tests the go accordion and panel components

* Adds dark theme to accordion panels

This allows panels to be displayed in dark mode for use on dark
backgrounds. It also updates the accordion to honor all of the
attributes explicitly set on the panel. This allows individual
panels to override the attributes defined by the accordion.

* Adjusts html semantics