Update input (#139)
* Refactors go-input component

This sets up go-input to accept a reactive form control and bases
much of the functionality off of that object. Using reactive forms
we can control most of the UI of a form input from the control.

This will allow us to handle all of the form input configuration
during implementation.