Fix button icon alignment (#176)
* Fixes issue with how the icon was aligned inside the button

View encapsulation was causing some of our button styles to not
be applied. This moves the styles out of the button component and
into the icon component so that they apply to the correct component.

* Clean up how classes are applied to both the go-button and go-icon

Previously, we used funcitons inside the template to declare the
classes on both the icon and button components. Switching this to
use lifecycle hooks and a simple object increases performance.

* Fixing tests that were failing due to functional changes

* Fixes potential issue with datepicker unsubscribe function

There are potential cases where the ngOnDestroy lifecycle hook
could be fired when a subscription doesn't exist. This causese the
unsubscribe function to fail and was throwing errors in our tests.
Checking for a subscription in those lifecycle hooks fixes the
issue.

* adds tests for the disabled input