Ability to choose which columns to sort by in table
Currently we have the ability to either have every column on a table sortable or none of them. There
are certain cases where we want to be able to just sort by a couple columns on a table. This provides
a way for an implementor to decide which columns can be sortable on a table or which columns they want
to not be sortable.