CHORE: Add Placeholder Binding to GoSelect
Some implementations of select boxes require a placeholder to be
designated for that select box. With our current version of the GoSelect
component, we are not capable of adding a placeholder value.

The NgSelect module allows us to pass it a placeholder for the select
box. This adds the placeholder option as a binding to the GoSelect
component and then leverages that binding when creating an NgSelect
instance.

I leverged the HTML for Template/Header/Footer[0] example that NgSelect
provides in order to best implement this solution.

[0] - https://github.com/ng-select/ng-select/blob/master/src/demo/app/examples/template-header-footer-example/template-header-footer-example.component.html