Manually trigger change detection on accordion panel after accordion is initialized.
We're updating the accordion panel from the accordion after it's initialized. The problem with
this is in production mode there isn't another lifecycle trigger after we change this and those changes
won't get picked up by the browser. This can be fixed by manually triggering the change detection cycle
after we update it so that the changes are picked up.