Reset Scroll Position on Page Navigation
When a user goes to a new page it should go to the top of the page instead of going to spot that they
were scrolled to on the previous page. If they are going back to the page they were just on, then it
should instead go to the spot they were previously at on the page.