Destroys dynamically created components when the off canvas and modal are closed
Previously the ngOnDestroy method would never get triggered for a component rendered inside of a modal or off canvas when they where closed. This creates issues when we need to run some cleanup code for whenever an off canvas or modal are closed because there is no way to know when they get closed.